changes to architecture,apis,static_path using moment
