+++ app/cart/routesDefinations/cart.routeDefs.js
@@ -1,0 +1,5 @@
+const cartRoute = require("../routes/cart.route");
+
+module.exports = function (app) {
+  app.use("/api/cart", cartRoute);
+};
