+++ app/product/routesDefinations/product.routeDefs.js
@@ -1,0 +1,5 @@
+const productRoute = require("../routes/product.route");
+
+module.exports = function (app) {
+  app.use("/api/product", productRoute);
+};
