+++ lib/config/db.js
@@ -1,7 +1,7 @@
 const mongoose = require("mongoose");
 const connectDB = async () => {
   const conn = await mongoose.connect(process.env.db);
-  console.log(`MongoDB Connected: ${conn.connection._connectionString}`);
+  console.log(`MongoDB Connected`);
 };
 
 module.exports = connectDB;
